[SMA-230] Refactor middleware
